Reorganize source code.
