--- .bazelrc
@@ -1,2 +1,0 @@
-build --incompatible_package_name_is_a_function=false
-query --incompatible_package_name_is_a_function=false
