+++ bazel/lightstep_build_system.bzl
@@ -87,6 +87,7 @@ def lightstep_cc_binary(
         srcs = [],
         data = [],
         testonly = 0,
+        linkshared = 0,
         visibility = None,
         external_deps = [],
         deps = [],
@@ -96,6 +97,7 @@ def lightstep_cc_binary(
         args = args,
         srcs = srcs,
         data = data,
+        linkshared = linkshared,
         copts = lightstep_include_copts() + lightstep_copts(),
         linkopts = linkopts,
         testonly = testonly,
