+++ src/common/fragment_array_input_stream.cpp
@@ -1,4 +1,4 @@
-#include "network/fragment_array_input_stream.h"
+#include "common/fragment_array_input_stream.h"
 
 #include <cassert>
 #include <iostream>
