+++ src/common/fragment_input_stream.h
