+++ src/recorder/stream_recorder/BUILD
@@ -39,7 +39,7 @@ lightstep_cc_library(
     ],
     deps = [
         "//src/common:chunk_circular_buffer_lib",
-        "//src/network:fragment_array_input_stream_lib",
+        "//src/common:fragment_array_input_stream_lib",
         ":utility_lib",
         ":fragment_span_input_stream_lib",
         ":stream_recorder_metrics_lib",
@@ -56,7 +56,7 @@ lightstep_cc_library(
     ],
     deps = [
         "//src/common:circular_buffer_lib",
-        "//src/network:fragment_array_input_stream_lib",
+        "//src/common:fragment_array_input_stream_lib",
         ":utility_lib",
     ],
 )
@@ -72,8 +72,8 @@ lightstep_cc_library(
     deps = [
         "//src/common:function_ref_lib",
         "//src/common:utility_lib",
-        "//src/network:fragment_input_stream_lib",
-        "//src/network:fragment_array_input_stream_lib",
+        "//src/common:fragment_input_stream_lib",
+        "//src/common:fragment_array_input_stream_lib",
         ":embedded_metrics_message_lib",
         ":span_stream_lib",
         ":stream_recorder_metrics_lib",
@@ -182,7 +182,7 @@ lightstep_cc_library(
         "//src/common:protobuf_lib",
         "//src/common:utility_lib",
         "//src/common:circular_buffer_lib",
-        "//src/network:fragment_array_input_stream_lib",
+        "//src/common:fragment_array_input_stream_lib",
     ],
 )
