+++ src/recorder/stream_recorder/connection_stream.h
@@ -6,8 +6,8 @@
 
 #include "common/function_ref.h"
 #include "common/utility.h"
-#include "network/fragment_array_input_stream.h"
-#include "network/fragment_input_stream.h"
+#include "common/fragment_array_input_stream.h"
+#include "common/fragment_input_stream.h"
 #include "recorder/stream_recorder/embedded_metrics_message.h"
 #include "recorder/stream_recorder/fragment_span_input_stream.h"
 #include "recorder/stream_recorder/span_stream.h"
