+++ test/common/BUILD
@@ -121,3 +121,26 @@ lightstep_catch_test(
       "//src/common:random_traverser_lib",
     ],
 )
+
+lightstep_catch_test(
+    name = "fragment_input_stream_test",
+    srcs = [
+        "fragment_input_stream_test.cpp",
+    ],
+    deps = [
+        "//src/common:fragment_array_input_stream_lib",
+        "//test:utility_lib",
+    ],
+)
+
+lightstep_catch_test(
+    name = "fragment_array_input_stream_test",
+    srcs = [
+        "fragment_array_input_stream_test.cpp",
+    ],
+    deps = [
+        "//src/common:fragment_array_input_stream_lib",
+        "//test:utility_lib",
+    ],
+)
+
