+++ test/common/fragment_array_input_stream_test.cpp
@@ -1,4 +1,4 @@
-#include "network/fragment_array_input_stream.h"
+#include "common/fragment_array_input_stream.h"
 
 #include <cstring>
 #include <string>
