+++ test/tracer/tracer-export-map.ld
@@ -1,0 +1,5 @@
+{
+  global:
+    OpenTracingMakeTracerFactory;
+  local: *;
+};
